Fixed issue with bmp stats causing a peer down to become up again.  Added docs for schema update to include prefix_bits.
IOS bmp stats are sent for a peer that is in either up or down state.
This caused a new peer message "first" to be sent, resulting in the
peer record being updated as UP.   The new behavior will now
only mark a peer as up if a PEER_UP message is received.
This means that JunOS <= 11.x peers will not show up because
it doesn't send peer up messages.   If this is a pain point, we can
update the code to derive peer up by ignoring stat updates for
a down peer.